Agent-Dev/requirements.txt
mem0ai==0.1.114
vecs==0.4.5
langgraph==0.3.30
langchain-core==0.3.67
langchain-community==0.3.26
langchain-openai==0.3.27
langchain-mcp-adapters==0.1.7
langgraph-cli[inmem]

